gene_network_quality_agent/requirements.txt
langgraph>=0.0.40
langchain>=0.1.0
langchain-openai>=0.1.0
pyyaml>=6.0
networkx>=3.0
openai>=1.0.0
python-dotenv>=1.0.0
